moved App.css into different directory for organization
