+++ src/Task.jsx
@@ -1,4 +1,4 @@
-import './App.css'
+import './css/App.css'
 
 function Task({ tasks = [], removeTask }) {
